# shared/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: SentinelVision API
  version: 0.1.0
  description: SyncCam AI external API retaining canonical wire identifiers pending a migration ADR.
servers:
  - url: https://api.sentinelvision.ai/v1
paths:
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Return the verified tenant-scoped user context
      security:
        - oidc: [auth:read]
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserEnvelope"
        "401": { description: Authentication required }
        "403": { description: Capability denied }
        "404": { description: Tenant scope not found }
  /sites:
    get:
      operationId: listSites
      summary: List sites visible inside the verified tenant and site scope
      security:
        - oidc: [sites:read]
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: Tenant- and site-scoped list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteListEnvelope"
        "401": { description: Authentication required }
        "403": { description: Capability denied }
        "404": { description: Tenant scope not found }
    post:
      operationId: createSite
      summary: Create one tenant site with an idempotent, audited mutation
      security:
        - oidc: [tenant:manage]
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteRequest"
      responses:
        "201":
          description: Site created or exact stored response replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteEnvelope"
        "401": { description: Authentication required }
        "403": { description: Capability denied }
        "404": { description: Tenant scope not found }
        "409": { description: Resource conflict or idempotency key reused with a different body }
        "422": { description: Request validation failed }
  /events:
    post:
      operationId: ingestEvent
      summary: Ingest a tenant-scoped, human-review event
      security:
        - oidc: [events:write]
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/EventRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectionEvent"
      responses:
        "202":
          description: Event accepted, or the same logical event replayed without another outbox message
          headers:
            Idempotent-Replayed:
              description: Present and true when an existing identical event is replayed.
              schema: { type: string, const: "true" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestEventEnvelope"
        "401":
          description: Authentication required
        "403":
          description: Tenant or capability denied
        "404":
          description: Tenant or site scope not found
        "409":
          description: Event identifier or dedupe key conflicts with different content
        "422":
          description: Event validation failed
        "503":
          description: Event persistence unavailable
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.sentinelvision.ai/.well-known/openid-configuration
  parameters:
    TenantHeader:
      name: X-SentinelVision-Tenant-ID
      in: header
      required: true
      description: Must exactly match the tenant_id claim in the verified bearer token.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Tenant-scoped mutation key retained for at least 24 hours.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Client UUIDv4 or a server-generated UUIDv4 returned in the response.
      schema:
        type: string
        format: uuid
    EventRequestId:
      name: X-SentinelVision-Request-ID
      in: header
      required: true
      description: UUIDv4 used to correlate the accepted event and append-only audit record.
      schema:
        type: string
        format: uuid
  schemas:
    DetectionEvent:
      type: object
      additionalProperties: false
      required:
        - event_id
        - tenant_id
        - dedupe_key
        - occurred_at
        - site_id
        - camera_id
        - zone_id
        - event_type
        - model_version
        - confidence
        - evidence_refs
        - requires_human_review
        - review_state
      properties:
        event_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        dedupe_key: { type: string, minLength: 1, maxLength: 256 }
        occurred_at: { type: string, format: date-time }
        site_id: { type: string, format: uuid }
        camera_id: { type: string, format: uuid }
        zone_id: { type: string, format: uuid }
        event_type:
          type: string
          enum: [camera_health, intrusion, restricted_zone, loitering, vehicle_activity, weapon_review, fire_review, smoke_review, attendance_review, ppe_review, fall_review, fight_review, abandoned_object_review]
        model_version: { type: string, minLength: 1 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        evidence_refs:
          type: array
          items: { type: string }
        requires_human_review:
          type: boolean
          const: true
        review_state:
          type: string
          enum: [pending, acknowledged, confirmed, dismissed]
    IngestEventEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [event_id, accepted]
          properties:
            event_id: { type: string, format: uuid }
            accepted: { type: boolean, const: true }
    CurrentUserEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CurrentUser"
    CurrentUser:
      type: object
      additionalProperties: false
      required: [id, tenant_id, site_ids, roles, scopes, data_class, mfa_level]
      properties:
        id: { type: string, minLength: 1 }
        email: { type: string, format: email }
        tenant_id: { type: string, minLength: 1, maxLength: 128 }
        site_ids:
          type: array
          items: { type: string, minLength: 1, maxLength: 128 }
        roles:
          type: array
          minItems: 1
          items:
            type: string
            enum: [super_admin, site_admin, operator, auditor, viewer]
        scopes:
          type: array
          items: { type: string, minLength: 1 }
        data_class:
          type: array
          items:
            type: string
            enum: [metadata, raw_video, biometric]
        mfa_level: { type: string, minLength: 1 }
    SiteListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Site" }
        meta:
          type: object
          additionalProperties: false
          required: [count, next]
          properties:
            count: { type: integer, minimum: 0 }
            next: { type: [string, "null"] }
    SiteEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/Site" }
    CreateSiteRequest:
      type: object
      additionalProperties: false
      required: [name, timezone]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        address: { type: string, maxLength: 500 }
        timezone: { type: string, minLength: 1, maxLength: 64 }
    Site:
      type: object
      additionalProperties: false
      required: [id, tenant_id, name, timezone, status, created_at]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        address: { type: string, maxLength: 500 }
        timezone: { type: string, minLength: 1, maxLength: 64 }
        status:
          type: string
          enum: [provisioning, active, offline, retired]
        created_at: { type: string, format: date-time }